requirements.txt
Flask
gunicorn
mysql-connector-python
bcrypt
smtplib
email